requirements.txt
tensorflow==2.10.0
keras==2.10.0
torch==1.13.1
scipy==1.9.3
matplotlib==3.6.2
scikit-learn==1.1.3
numpy==1.23.5
torchvision==0.14.1
pillow==11.1.0
pandas==2.2.3
tqdm==4.67.1